+++ includes/blocks/jb-job-post/package.json
@@ -1,0 +1,11 @@
+{
+  "scripts": {
+    "build": "wp-scripts build",
+    "start": "wp-scripts start"
+  },
+  "dependencies": {
+    "@wordpress/block-editor": "^11.5.0",
+    "@wordpress/components": "^23.5.0",
+    "@wordpress/i18n": "^4.28.0"
+  }
+}
