+++ includes/blocks/jb-job-post/src/index.js
@@ -1,0 +1,22 @@
+import { registerBlockType } from '@wordpress/blocks';
+import ServerSideRender from '@wordpress/server-side-render';
+import { useBlockProps } from '@wordpress/block-editor';
+
+registerBlockType('jb-block/jb-job-post', {
+	edit: function (props) {
+		jQuery('#jb-job-preview, #jb-job-draft, #jb_company_logo_plupload').attr('disabled', 'disabled');
+		const blockProps = useBlockProps();
+
+		return (
+			<div {...blockProps}>
+				<ServerSideRender block="jb-block/jb-job-post" />
+			</div>
+		);
+
+	},
+
+	save: function (props) {
+		return null;
+	}
+
+});
