+++ includes/blocks/jb-jobs-categories-list/block.json
@@ -5,5 +5,5 @@
   "name": "jb-block/jb-jobs-categories-list",
   "category": "jb-blocks",
   "icon": "editor-ul",
-  "editorScript": "file:./block.js"
+  "editorScript": "file:./build/index.js"
 }
