+++ includes/blocks/jb-jobs-categories-list/package.json
@@ -1,0 +1,16 @@
+{
+  "scripts": {
+    "build": "wp-scripts build",
+    "start": "wp-scripts start"
+  },
+  "dependencies": {
+    "@wordpress/block-editor": "^11.5.0",
+    "@wordpress/components": "^23.5.0",
+    "@wordpress/i18n": "^4.28.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
+  },
+  "devDependencies": {
+    "@wordpress/server-side-render": "^4.5.0"
+  }
+}
